backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
groq==0.4.1
httpx==0.27.0
twilio==8.11.1
apscheduler==3.10.4
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-multipart==0.0.6
httpcore==1.0.5
